using common middleware for all http servers
