--- nym-credential-proxy/nym-credential-proxy/src/http/middleware/mod.rs
@@ -1,5 +1,0 @@
-// Copyright 2024 Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
-
-pub mod auth;
-pub mod logging;
