+++ nym-credential-proxy/nym-credential-proxy/src/http/router/api/v1/mod.rs
@@ -1,13 +1,11 @@
 // Copyright 2024 Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::http::middleware::auth::AuthLayer;
 use crate::http::state::ApiState;
 use axum::Router;
 use nym_credential_proxy_requests::routes::api::v1;
+use nym_http_api_common::middleware::bearer_auth::AuthLayer;
 
-// pub mod bandwidth_voucher;
-// pub mod freepass;
 pub mod openapi;
 pub mod ticketbook;
