--- nym-node/nym-node-http-api/src/middleware/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
-
-pub mod logging;
